Moved some code around between node_context_windows.py and context_windows.py
